Move ComparisonGraphs to separate component. Update loading logic to check for each property indivdually
